+++ lib/packages/wikibase/data-model/src/HashArray.php
@@ -4,6 +4,7 @@ namespace Wikibase\DataModel;
 
 use Hashable;
 use InvalidArgumentException;
+use Wikibase\DataModel\Internal\MapValueHasher;
 
 /**
  * Generic array object with lookups based on hashes of the elements.
@@ -281,8 +282,6 @@ abstract class HashArray extends \ArrayObject implements \Hashable, \Comparable 
 	 * The hash is purely valuer based. Order of the elements in the array is not held into account.
 	 *
 	 * @since 0.1
-	 *
-	 * @internal param MapHasher $mapHasher
 	 *
 	 * @return string
 	 */