+++ lib/packages/wikibase/data-model/RELEASE-NOTES.md
@@ -48,21 +48,24 @@ Other breaking changes:
 * Calling `clear` on an `Item` will now cause its statements to be removed
 * `SiteLinkList::addNewSiteLink` no longer returns a `SiteLinkList` instance
 * Removed the global variable `evilDataValueMap`
+* Removed `ClaimAggregate` interface, which is thus no longer implemented by `Entity`
 
 #### Additions
 
-* Added `getClaim` method to `Statement`
 * Added `ClaimList`
 * Added `ClaimListDiffer`
 * Added `PropertyDataTypeLookup` and trivial implementation `InMemoryDataTypeLookup`
 * Added `PropertyNotFoundException`
 * Added `ItemDiffer` and `PropertyDiffer`
 * Added `EntityDiffer` and `EntityDifferStrategy`
+* Added `Item::getStatements`
+* Added `Statement::getClaim`
 
 #### Other changes
 
 * Undeprecated passing an integer to `Item::setId` and `Property::setId`
 * Deprecated `Entity::setId`, `Entity::newClaim` and `Entity::getAllSnaks`
+* Deprecated `Item::getClaims` in favour of `Item::getStatements`
 
 ## Version 0.9 (2014-08-15)